chore: added gesture handler root view
