+++ example/app/src/components/contactList/styles.ts
@@ -1,0 +1,21 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+  sectionHeaderContainer: {
+    paddingTop: 24,
+    paddingBottom: 6,
+    backgroundColor: 'white',
+  },
+  sectionHeaderTitle: {
+    fontSize: 16,
+    textTransform: 'uppercase',
+  },
+  container: {
+    overflow: 'visible',
+    flex: 1,
+  },
+  contentContainer: {
+    paddingHorizontal: 16,
+    overflow: 'visible',
+  },
+});
