+++ example/app/src/components/contactList/styles.web.ts
@@ -1,0 +1,20 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+  sectionHeaderContainer: {
+    paddingTop: 24,
+    paddingBottom: 6,
+    backgroundColor: 'white',
+  },
+  sectionHeaderTitle: {
+    fontSize: 16,
+    textTransform: 'uppercase',
+  },
+  container: {
+    flex: 1,
+    paddingHorizontal: 16,
+  },
+  contentContainer: {
+    paddingHorizontal: 16,
+  },
+});
