+++ example/app/src/index.ts
@@ -5,3 +5,4 @@ export { Button } from './components/button';
 export { ContactList } from './components/contactList';
 export { ContactItem } from './components/contactItem';
 export { SearchHandle, SEARCH_HANDLE_HEIGHT } from './components/searchHandle';
+export * from './utilities/createMockData';
