+++ example/bare/package.json
@@ -24,7 +24,7 @@
     "nanoid": "^3.3.3",
     "react": "17.0.2",
     "react-native": "0.68.1",
-    "react-native-gesture-handler": "^2.1.0",
+    "react-native-gesture-handler": "^2.4.2",
     "react-native-maps": "^0.30.1",
     "react-native-pager-view": "^5.4.9",
     "react-native-reanimated": "^2.8.0",
