+++ example/expo/src/App.tsx
@@ -1,17 +1,4 @@
 import React from 'react';
-// import { screens } from './screens';
 import { App } from '@gorhom/bottom-sheet-example-app';
-import { GestureHandlerRootView } from 'react-native-gesture-handler';
-import { StyleSheet } from 'react-native';
 
-export default () => (
-  <GestureHandlerRootView style={styles.container}>
-    <App />
-  </GestureHandlerRootView>
-);
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-});
+export default () => <App />;
